chore(parquet): Make parquet WASM loader into the primary loader
