+++ modules/parquet/src/index.ts
@@ -6,22 +6,24 @@ export {Buffer} from './polyfills/buffer/install-buffer-polyfill';
 
 // import {ArrowTable, ArrowTableBatch} from '@loaders.gl/arrow';
 
+export {ParquetFormat} from './parquet-format';
+
 export {
-  ParquetWorkerLoader,
-  ParquetLoader,
+  ParquetJSONWorkerLoader,
+  ParquetJSONLoader,
   GeoParquetWorkerLoader,
   GeoParquetLoader,
-  ParquetColumnarWorkerLoader,
-  ParquetColumnarLoader
-} from './parquet-loader';
+  // ParquetJSONColumnarWorkerLoader,
+  // ParquetJSONColumnarLoader
+} from './parquet-json-loader';
 
-export {ParquetWriter as _ParquetWriter} from './parquet-writer';
+export {ParquetJSONWriter as _ParquetJSONWriter} from './parquet-json-writer';
 
 // EXPERIMENTAL - expose Parquet WASM loaders/writer
 
-export type {ParquetWasmLoaderOptions} from './parquet-wasm-loader';
-export {ParquetWasmLoader, ParquetWasmWorkerLoader} from './parquet-wasm-loader';
-export {ParquetWasmWriter} from './parquet-wasm-writer';
+export type {ParquetArrowLoaderOptions} from './parquet-arrow-loader';
+export {ParquetArrowLoader, ParquetArrowWorkerLoader} from './parquet-arrow-loader';
+export {ParquetArrowWriter} from './parquet-arrow-writer';
 
 // EXPERIMENTAL - expose the internal parquetjs API
