+++ modules/parquet/src/lib/parsers/get-parquet-schema.ts
@@ -4,9 +4,9 @@
 
 // loaders.gl
 import {Schema} from '@loaders.gl/schema';
+import {unpackGeoMetadata, unpackJSONStringMetadata} from '@loaders.gl/gis';
 import {ParquetReader} from '../../parquetjs/parser/parquet-reader';
 import {convertParquetSchema} from '../arrow/convert-schema-from-parquet';
-import {unpackGeoMetadata, unpackJSONStringMetadata} from '@loaders.gl/gis';
 
 export async function getSchemaFromParquetReader(reader: ParquetReader): Promise<Schema> {
   const parquetSchema = await reader.getSchema();
