+++ modules/parquet/src/parquet-format.ts
@@ -1,0 +1,14 @@
+// loaders.gl
+// SPDX-License-Identifier: MIT
+// Copyright (c) vis.gl contributors
+
+import type {Format} from '@loaders.gl/loader-utils';
+
+export const ParquetFormat = {
+  name: 'Apache Parquet',
+  category: 'table',
+  extensions: ['parquet'],
+  mimeTypes: ['application/octet-stream'],
+  binary: true,
+  tests: ['PAR1', 'PARE'],
+} as const satisfies Partial<Format>;
