+++ modules/parquet/src/parquet-json-loader.ts
@@ -8,17 +8,21 @@ import type {
   ObjectRowTableBatch,
   GeoJSONTable,
   GeoJSONTableBatch,
-  ColumnarTable,
-  ColumnarTableBatch
+  // ColumnarTable,
+  // ColumnarTableBatch
 } from '@loaders.gl/schema';
 import {BlobFile} from '@loaders.gl/loader-utils';
 
-import {parseParquetFile, parseParquetFileInBatches} from './lib/parsers/parse-parquet';
-import {parseGeoParquetFile, parseGeoParquetFileInBatches} from './lib/parsers/parse-geoparquet';
+import {parseParquetFile, parseParquetFileInBatches} from './lib/parsers/parse-parquet-to-json';
 import {
-  parseParquetFileInColumns,
-  parseParquetFileInColumnarBatches
-} from './lib/parsers/parse-parquet-to-columns';
+  parseGeoParquetFile,
+  parseGeoParquetFileInBatches
+} from './lib/parsers/parse-geoparquet-to-geojson';
+// import {
+//   parseParquetFileInColumns,
+//   parseParquetFileInColumnarBatches
+// } from './lib/parsers/parse-parquet-to-columns';
+import { ParquetFormat } from './parquet-format';
 
 // Note: The Buffer polyfill is quite fragile
 // For some reason, just exporting directly fails with some bundlers
@@ -31,11 +35,9 @@ export {Buffer};
 const VERSION = typeof __VERSION__ !== 'undefined' ? __VERSION__ : 'latest';
 
 /** Options for the parquet loader */
-export type ParquetLoaderOptions = LoaderOptions & {
+export type ParquetJSONLoaderOptions = LoaderOptions & {
   /** Options for the parquet loader */
   parquet?: {
-    /** Format of returned parsed data */
-    shape?: 'object-row-table' | 'geojson-table';
     /** Restrict which columns that are parsed from the table. Can save significant memory. */
     columnList?: string[] | string[][];
     /** If true, binary values are not converted to strings */
@@ -50,47 +52,42 @@ export type ParquetLoaderOptions = LoaderOptions & {
 /**
  * ParquetJS table loader
  */
-export const ParquetWorkerLoader = {
+export const ParquetJSONWorkerLoader = {
+  ...ParquetFormat,
+  
   dataType: null as unknown as ObjectRowTable,
   batchType: null as unknown as ObjectRowTableBatch,
 
-  name: 'Apache Parquet',
   id: 'parquet',
   module: 'parquet',
   version: VERSION,
   worker: false,
-  category: 'table',
-  extensions: ['parquet'],
-  mimeTypes: ['application/octet-stream'],
-  binary: true,
-  tests: ['PAR1', 'PARE'],
   options: {
     parquet: {
-      shape: 'object-row-table',
       columnList: [],
       geoparquet: true,
       url: undefined,
       preserveBinary: false
     }
   }
-} as const satisfies Loader<ObjectRowTable, ObjectRowTableBatch, ParquetLoaderOptions>;
+} as const satisfies Loader<ObjectRowTable, ObjectRowTableBatch, ParquetJSONLoaderOptions>;
 
 /** ParquetJS table loader */
-export const ParquetLoader = {
-  ...ParquetWorkerLoader,
+export const ParquetJSONLoader = {
+  ...ParquetJSONWorkerLoader,
 
-  dataType: null as unknown as ObjectRowTable | GeoJSONTable,
-  batchType: null as unknown as ObjectRowTableBatch | GeoJSONTableBatch,
+  dataType: null as unknown as ObjectRowTable,
+  batchType: null as unknown as ObjectRowTableBatch,
 
-  parse: (arrayBuffer: ArrayBuffer, options?: ParquetLoaderOptions) =>
+  parse: (arrayBuffer: ArrayBuffer, options?: ParquetJSONLoaderOptions) =>
     parseParquetFile(new BlobFile(arrayBuffer), options),
 
   parseFile: parseParquetFile,
   parseFileInBatches: parseParquetFileInBatches
 } as const satisfies LoaderWithParser<
-  ObjectRowTable | GeoJSONTable,
-  ObjectRowTableBatch | GeoJSONTableBatch,
-  ParquetLoaderOptions
+  ObjectRowTable,
+  ObjectRowTableBatch,
+  ParquetJSONLoaderOptions
 >;
 
 // Defeat tree shaking
@@ -98,35 +95,31 @@ export const ParquetLoader = {
 ParquetLoader.Buffer = Buffer;
 
 export const GeoParquetWorkerLoader = {
+  ...ParquetFormat,
+
   dataType: null as unknown as GeoJSONTable,
   batchType: null as unknown as GeoJSONTableBatch,
 
-  name: 'Apache Parquet',
   id: 'parquet',
   module: 'parquet',
   version: VERSION,
   worker: true,
-  category: 'table',
-  extensions: ['parquet'],
-  mimeTypes: ['application/octet-stream'],
-  binary: true,
-  tests: ['PAR1', 'PARE'],
+
   options: {
     parquet: {
-      shape: 'geojson-table',
       columnList: [],
       geoparquet: true,
       url: undefined,
       preserveBinary: false
     }
   }
-} as const satisfies Loader<GeoJSONTable, GeoJSONTableBatch, ParquetLoaderOptions>;
+} as const satisfies Loader<GeoJSONTable, GeoJSONTableBatch, ParquetJSONLoaderOptions>;
 
 /** ParquetJS table loader */
 export const GeoParquetLoader = {
   ...GeoParquetWorkerLoader,
 
-  parse(arrayBuffer: ArrayBuffer, options?: ParquetLoaderOptions) {
+  parse(arrayBuffer: ArrayBuffer, options?: ParquetJSONLoaderOptions) {
     return parseGeoParquetFile(new BlobFile(arrayBuffer), options);
   },
   parseFile: parseGeoParquetFile,
@@ -134,11 +127,11 @@ export const GeoParquetLoader = {
 } as const satisfies LoaderWithParser<
   ObjectRowTable | GeoJSONTable,
   ObjectRowTableBatch | GeoJSONTableBatch,
-  ParquetLoaderOptions
+  ParquetJSONLoaderOptions
 >;
 
-/** @deprecated Test to see if we can improve perf of parquetjs loader */
-export const ParquetColumnarWorkerLoader = {
+/** @deprecated Test to see if we can improve perf of parquetjs loader *
+export const ParquetJSONColumnarWorkerLoader = {
   dataType: null as any as ColumnarTable,
   batchType: null as any as ColumnarTableBatch,
 
@@ -152,15 +145,16 @@ export const ParquetColumnarWorkerLoader = {
   mimeTypes: ['application/octet-stream'],
   binary: true,
   tests: ['PAR1', 'PARE'],
-  options: ParquetLoader.options
-} as const satisfies Loader<ColumnarTable, ColumnarTableBatch, ParquetLoaderOptions>;
+  options: ParquetJSONLoader.options
+} as const satisfies Loader<ColumnarTable, ColumnarTableBatch, ParquetJSONLoaderOptions>;
 
-/** @deprecated Test to see if we can improve perf of parquetjs loader */
-export const ParquetColumnarLoader = {
-  ...ParquetColumnarWorkerLoader,
-  parse(arrayBuffer: ArrayBuffer, options?: ParquetLoaderOptions) {
+/** @deprecated Test to see if we can improve perf of parquetjs loader *
+export const ParquetJSONColumnarLoader = {
+  ...ParquetJSONColumnarWorkerLoader,
+  parse(arrayBuffer: ArrayBuffer, options?: ParquetJSONLoaderOptions) {
     return parseParquetFileInColumns(new BlobFile(arrayBuffer), options);
   },
   parseFile: parseParquetFileInColumns,
   parseFileInBatches: parseParquetFileInColumnarBatches
-} as const satisfies LoaderWithParser<ColumnarTable, ColumnarTableBatch, ParquetLoaderOptions>;
+} as const satisfies LoaderWithParser<ColumnarTable, ColumnarTableBatch, ParquetJSONLoaderOptions>;
+*/
